make solver_result and solution_result separate targets
