+++ solvers/BUILD.bazel
@@ -48,7 +48,9 @@ drake_cc_package_library(
         ":rotation_constraint",
         ":scs_solver",
         ":snopt_solver",
+        ":solution_result",
         ":solver_id",
+        ":solver_result",
         ":solver_type",
         ":solver_type_converter",
         ":sos_basis_generator",
@@ -262,7 +264,9 @@ drake_cc_library(
         ":osqp_solver",
         ":scs_solver",
         ":snopt_solver",
+        ":solution_result",
         ":solver_id",
+        ":solver_result",
         ":sos_basis_generator",
         ":symbolic_extraction",
         ":unrevised_lemke_solver",
@@ -278,7 +282,8 @@ drake_cc_library(
     srcs = ["mathematical_program_result.cc"],
     hdrs = ["mathematical_program_result.h"],
     deps = [
-        ":mathematical_program_api",
+        ":solution_result",
+        ":solver_result",
         "//systems/framework:value",
     ],
 )
@@ -488,6 +493,24 @@ drake_cc_library(
     ],
 )
 
+drake_cc_library(
+    name = "solution_result",
+    srcs = [
+        "solution_result.cc",
+    ],
+    hdrs = ["solution_result.h"],
+)
+
+drake_cc_library(
+    name = "solver_result",
+    srcs = ["solver_result.cc"],
+    hdrs = ["solver_result.h"],
+    deps = [
+        ":solver_id",
+        "//common:essential",
+    ],
+)
+
 # The mathematical_program_api target exists so that solver targets can access
 # the mathematical_program.h header, and the :mathematical_program target can
 # depend on solver targets, without creating a circular dependency.
@@ -512,7 +535,9 @@ drake_cc_library(
         ":decision_variable",
         ":function",
         ":indeterminate",
+        ":solution_result",
         ":solver_id",
+        ":solver_result",
         "//common:autodiff",
         "//common:essential",
         "//common:polynomial",
